Add unit test, move sanitize to MetaMorpheusTask
