+++ Test/AmbiguityTest.cs
@@ -1,3 +1,4 @@
+using Chemistry;
 using EngineLayer;
 using EngineLayer.ClassicSearch;
 using MzLibUtil;
@@ -139,6 +140,60 @@ namespace Test
 
 
             Directory.Delete(outputFolder, true);
+        }
+
+        public class SanitizeTestClass : SearchTask
+        {
+            [Test]
+            public static void TestDatabaseSanitizationPrioritizesXml()
+            {
+                //if a fasta and a xml database are both used and called targets, prioritize the xml
+                Protein fastaProtein = new Protein("APEPTIDE", "Test");
+                ModificationMotif.TryGetMotif("A", out ModificationMotif motif);
+                Modification mod = new Modification(_originalId: "acetylation", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("C2H2O1"), _locationRestriction: "Anywhere.");
+                Protein xmlProtein = new Protein("APEPTIDE", "Test", oneBasedModifications: new Dictionary<int, List<Modification>> { { 1, new List<Modification> { mod } } });
+                List<Protein> proteins = new List<Protein> { fastaProtein, xmlProtein };
+                SanitizeProteinDatabase(proteins, TargetContaminantAmbiguity.RemoveTarget);
+                Assert.IsTrue(proteins.Count == 1);
+                Assert.IsTrue(proteins.First().OneBasedPossibleLocalizedModifications.Count != 0);
+
+                //reverse order and try again
+                proteins = new List<Protein> { xmlProtein, fastaProtein };
+                SanitizeProteinDatabase(proteins, TargetContaminantAmbiguity.RemoveTarget);
+                Assert.IsTrue(proteins.Count == 1);
+                Assert.IsTrue(proteins.First().OneBasedPossibleLocalizedModifications.Count != 0);
+
+                //same with no mods
+                xmlProtein = new Protein("APEPTIDE", "Test", proteolysisProducts: new List<ProteolysisProduct> { new ProteolysisProduct(1, 3, "zrWuzHere") });
+                proteins = new List<Protein> { xmlProtein, fastaProtein };
+                SanitizeProteinDatabase(proteins, TargetContaminantAmbiguity.RemoveTarget);
+                Assert.IsTrue(proteins.Count == 1);
+                Assert.IsTrue(proteins.First().ProteolysisProducts.Count() != 0);
+            }
+
+            [Test]
+            public static void TestDatabaseSanitizationRemovesDecoys()
+            {
+                ModificationMotif.TryGetMotif("A", out ModificationMotif motif);
+                Modification mod = new Modification(_originalId: "acetylation", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("C2H2O1"), _locationRestriction: "Anywhere.");
+                Protein targetProtein = new Protein("APEPTIDE", "Test", oneBasedModifications: new Dictionary<int, List<Modification>> { { 1, new List<Modification> { mod } } });
+                Protein contaminantProtein = new Protein("APEPTIDE", "Test", isContaminant: true);
+                Protein tDecoyProtein = new Protein("AEDITPEP", "DECOY_Test", oneBasedModifications: new Dictionary<int, List<Modification>> { { 1, new List<Modification> { mod } } });
+                Protein cDecoyProtein = new Protein("AEDITPEP", "DECOY_Test");
+                List<Protein> proteins = new List<Protein> { targetProtein, contaminantProtein, tDecoyProtein, cDecoyProtein }; //two decoys, one for target, one for contaminant
+                SanitizeProteinDatabase(proteins, TargetContaminantAmbiguity.RemoveContaminant);
+                Assert.IsTrue(proteins.Count == 2);
+                Assert.IsTrue(!proteins[1].IsContaminant); //forward is target
+                Assert.IsTrue(proteins[0].OneBasedPossibleLocalizedModifications.Count == 1);
+                Assert.IsTrue(proteins[1].OneBasedPossibleLocalizedModifications.Count == 1);
+
+                proteins = new List<Protein> { targetProtein, contaminantProtein, tDecoyProtein, cDecoyProtein }; //two decoys, one for target, one for contaminant
+                SanitizeProteinDatabase(proteins, TargetContaminantAmbiguity.RemoveTarget);
+                Assert.IsTrue(proteins.Count == 2);
+                Assert.IsTrue(proteins[1].IsContaminant); //forward is contaminant
+                Assert.IsTrue(proteins[0].OneBasedPossibleLocalizedModifications.Count == 1); //would be cool if this was 0 and we could keep the correct decoy, but kind of an edge case
+                Assert.IsTrue(proteins[1].OneBasedPossibleLocalizedModifications.Count == 0); //contaminant doesn't have mods
+            }
         }
     }
 }
